Move ServiceLifecycle to :tooling-api (the only place it's used in
